scatter 4M

Created-by: maxinyuan1
Commit-by: maxinyuan1
Merged-by: turing_project1
Description: scatter 4M

See merge request: cann/hcomm-dev!146

+++ src/algorithm/impl/resource_manager/ccl_buffer_manager.h
@@ -57,7 +57,7 @@ public:
     void ReleaseAlltoAllvParaBuffer();
     HcclResult CleanCCLbuffer();
     HcclResult CleanAIVbuffer(void *bufferPtr);
-    HcclResult GetIndependentOpCCLbuffer(void* &buffer, u32 &size);
+    HcclResult GetIndependentOpCCLbuffer(void* &buffer, uint64_t &size);
 private:
     HcclResult CreateCCLbuffer(u64 size, DeviceMem &buffer);
     void* GetCCLbufferAddr(const DeviceMem &buffer);
